+++ bpf/libbpf/tracepoint.c
@@ -59,6 +59,9 @@ void handle_event(void *ctx, int cpu, void *event, unsigned int event_sz)
 
 	printf("%-6d %-6d %-16s(%-16s) %s ret = %d\n", m->pid, m->uid, m->comm,
 		m->comm2, m->filename, m->ret);
+#ifdef PARSE_AUXV
+	printf(" auxv.type %ld, auxv.val 0x%lx\n", m->auxv.type, m->auxv.val);
+#endif
 }
 
 void lost_event(void *ctx, int cpu, long long unsigned int event_sz)